The once described as the Pearl of Africa
The cradle of the most hospitable naturally endowed people, rich in culture and the flourishing flora and fauna are now a myth, her foundations for a brighter future over years post independence under-looked by the selfish tyrants and despots presumed to be our saviours, yet disillusioned with extortion and theft of our natural resources to the benefit of their families. leaving the entire population languishing in absolute poverty.